egs3/diar_sortformer_cache: self-contained Sortformer recipe (move + rename)
Rename the recipe egs3/librispeech_sortformer/diar -> egs3/diar_sortformer_cache
and make it self-contained: move all Sortformer-specific code out of the shared
library into the recipe's `sortformer/` package, so the recipe carries its own
model/encoders/cache modules/loss/sliding-window/longform/converters, task,
DiarizationSystem, DER metric, and inference bin.

- espnet2/diar/sortformer/* + espnet2/diar/espnet_sortformer_model.py
  + espnet2/tasks/diar_sortformer.py + espnet2/bin/diar_sortformer_inference.py
  + espnet3/systems/diar/{system,task,metrics/der}.py
  -> egs3/diar_sortformer_cache/sortformer/ (relative imports within the package).
- test -> egs3/diar_sortformer_cache/tests/test_sortformer.py.
- conf task:/_target_ and run.py re-point to sortformer.{task,system,der,
  inference_bin}; path.sh fixed for the shallower path; recipe .gitignore added.

Nothing outside the recipe referenced these. Verified: package imports, model
build via sortformer.task, and the test suite all pass; the shared espnet3
framework (ASRSystem/BaseSystem) is untouched.
