+++ egs3/diar_sortformer_cache/.gitignore
@@ -1,0 +1,5 @@
+data/
+exp/
+downloads/
+__pycache__/
+*.pyc
