+++ egs3/diar_sortformer_cache/conf/inference.yaml
@@ -9,7 +9,7 @@
 #
 # `model_file` defaults to the best trained checkpoint. To evaluate NVIDIA's
 # converted pretrained weights instead, point it at the .pth produced by
-#   python -m espnet2.diar.sortformer.convert_hf_sortformer \
+#   python -m sortformer.convert_hf_sortformer \
 #       --hf_model nvidia/diar_sortformer_4spk-v1 --out exp/sortformer_4spk.pth
 
 dataset:
@@ -32,7 +32,7 @@ parallel:
 #                    MODEL AND OUTPUT                    #
 ##########################################################
 model:
-  _target_: espnet2.bin.diar_sortformer_inference.SortformerDiarization
+  _target_: sortformer.inference_bin.SortformerDiarization
   train_config: ${exp_dir}/config.yaml
   model_file: ${exp_dir}/last.ckpt
