+++ egs3/diar_sortformer_cache/path.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+export PYTHONPATH=../../:../TEMPLATE/asr:$(pwd):${PYTHONPATH:-}
+
+source ../../tools/activate_python.sh
+source ../../tools/extra_path.sh
