+++ egs3/diar_sortformer_cache/sortformer/task.py
@@ -16,17 +16,18 @@ import numpy as np
 import torch
 from typeguard import typechecked
 
-from espnet2.diar.espnet_sortformer_model import ESPnetSortformerModel
-from espnet2.diar.sortformer.fastconformer_encoder import FastConformerEncoder
-from espnet2.diar.sortformer.preprocessor import MelSpectrogramPreprocessor
-from espnet2.diar.sortformer.sortformer_modules import SortformerModules
-from espnet2.diar.sortformer.transformer_encoder import TransformerEncoder
 from espnet2.tasks.abs_task import AbsTask
 from espnet2.torch_utils.initialize import initialize
 from espnet2.train.collate_fn import CommonCollateFn
 from espnet2.train.trainer import Trainer
 from espnet2.utils.nested_dict_action import NestedDictAction
 from espnet2.utils.types import int_or_none, str2bool, str_or_none
+
+from .fastconformer_encoder import FastConformerEncoder
+from .model import ESPnetSortformerModel
+from .preprocessor import MelSpectrogramPreprocessor
+from .sortformer_modules import SortformerModules
+from .transformer_encoder import TransformerEncoder
 
 
 class SortformerDiarizationTask(AbsTask):
@@ -182,7 +183,7 @@ class SortformerDiarizationTask(AbsTask):
         # Initialize the FastConformer from NEST self-supervised weights.
         init_nest = getattr(args, "init_nest", None)
         if init_nest is not None:
-            from espnet2.diar.sortformer.convert_nest import load_nest_encoder
+            from .convert_nest import load_nest_encoder
 
             load_nest_encoder(model, init_nest)
         return model
