--- egs3/librispeech_sortformer/diar/path.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-export PYTHONPATH=../../../:../../TEMPLATE/asr:$(pwd):${PYTHONPATH:-}
-
-source ../../../tools/activate_python.sh
-source ../../../tools/extra_path.sh
