--- espnet3/systems/diar/__init__.py
@@ -1,1 +1,0 @@
-"""Diarization system package for ESPnet3 (Sortformer)."""
