--- espnet3/systems/diar/metrics/__init__.py
@@ -1,1 +1,0 @@
-"""Diarization metric implementations (DER)."""
